refactor: use Context API instead of props chaining in Header component
