+++ src/app/Footer/Footer.js
@@ -1,6 +1,6 @@
-import Link from "next/link";
+import Link from 'next/link';
 
-import "./Footer.css";
+import './Footer.css';
 
 export default function Footer() {
   return (
@@ -14,19 +14,19 @@ export default function Footer() {
             </Link>
           </li>
           <li>
-            {" "}
+            {' '}
             <Link className="footer-nav-item" href="/About">
               About Us
             </Link>
           </li>
           <li>
-            {" "}
+            {' '}
             <Link className="footer-nav-item" href="/Contact">
               Contact
             </Link>
           </li>
           <li>
-            {" "}
+            {' '}
             <Link className="footer-nav-item" href="/Assignment-3">
               Assignment-3
             </Link>
